xamila_backend/requirements.txt
# ================================
# XAMILA BACKEND - REQUIREMENTS
# Plateforme Fintech Unifiée avec Matching SGI
# ================================

# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Database
mysqlclient==2.2.0
redis==5.0.1
django-redis==5.4.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7

# Email & Notifications
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# File Handling & Storage
Pillow==10.1.0
django-storages==1.14.2
boto3==1.29.7

# API Documentation
drf-spectacular==0.26.5
drf-spectacular[sidecar]==0.26.5

# Utilities
python-decouple==3.8
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Monitoring & Logging
sentry-sdk==1.38.0
django-health-check==3.17.0

# SGI Matching & Analytics
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2

# Mobile Money Integrations
pycryptodome==3.19.0
